rename fakemetrics -> mt-fakemetrics + associated fixes
