+++ .gitignore
@@ -4,10 +4,10 @@
 .vscode
 /build
 /cmd-dev/mt-simulate-memory-idx-lock-contention/mt-simulate-memory-idx-lock-contention
-/cmd/fakemetrics/fakemetrics
 /cmd/metrictank/metrictank
 /cmd/mt-aggs-explain/mt-aggs-explain
 /cmd/mt-explain/mt-explain
+/cmd/mt-fakemetrics/mt-fakemetrics
 /cmd/mt-gateway/mt-gateway
 /cmd/mt-index-cat/mt-index-cat
 /cmd/mt-index-migrate/mt-index-migrate
