+++ cmd/mt-fakemetrics/cmd/agginput.go
