+++ cmd/mt-fakemetrics/cmd/bad.go
@@ -17,7 +17,7 @@ package cmd
 import (
 	"time"
 
-	"github.com/grafana/metrictank/cmd/fakemetrics/out"
+	"github.com/grafana/metrictank/cmd/mt-fakemetrics/out"
 	"github.com/grafana/metrictank/schema"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
