+++ cmd/mt-fakemetrics/cmd/rand.go
