+++ cmd/mt-fakemetrics/cmd/resolutionchange.go
@@ -19,18 +19,18 @@ import (
 
 	"time"
 
-	"github.com/grafana/metrictank/cmd/fakemetrics/out"
+	"github.com/grafana/metrictank/cmd/mt-fakemetrics/out"
 	"github.com/grafana/metrictank/schema"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
-// storageconfCmd represents the storageconf command
-var storageconfCmd = &cobra.Command{
-	Use:   "storageconf",
-	Short: "Sends out one or more set of 10 metrics which you can test aggregation and retention rules on",
+// resolutionchangeCmd represents the resolutionchange command
+var resolutionchangeCmd = &cobra.Command{
+	Use:   "resolutionchange",
+	Short: "Sends out metric with changing intervals, time range 24hours",
 	Run: func(cmd *cobra.Command, args []string) {
-		initStats(true, "storageconf")
+		initStats(true, "resolutionchange")
 		if mpo%10 != 0 {
 			log.Fatal("mpo must divide by 10")
 		}
@@ -43,37 +43,22 @@ var storageconfCmd = &cobra.Command{
 		to := time.Now().Unix()
 		from := to - 24*60*60
 
-		names := []string{
-			"fakemetrics.raw.%d.min",
-			"fakemetrics.raw.%d.max",
-			"fakemetrics.raw.%d.sum",
-			"fakemetrics.raw.%d.lst",
-			"fakemetrics.raw.%d.default",
-			"fakemetrics.agg.%d.min",
-			"fakemetrics.agg.%d.max",
-			"fakemetrics.agg.%d.sum",
-			"fakemetrics.agg.%d.lst",
-			"fakemetrics.agg.%d.default",
-		}
 		var metrics []*schema.MetricData
-		sets := mpo / 10
-		for set := 1; set <= sets; set++ {
-			for _, name := range names {
-				metrics = append(metrics, buildMetric(name, set, 1))
-			}
+		for i := 0; i < mpo; i++ {
+			metrics = append(metrics, buildResChangeMetric("mt-fakemetrics.reschange.%d", i, 1))
 		}
-		do(metrics, from, to, outs)
+		runResolutionChange(metrics, from, to, outs)
 	},
 }
 
 func init() {
-	rootCmd.AddCommand(storageconfCmd)
-	storageconfCmd.Flags().IntVar(&mpo, "mpo", 10, "how many metrics per org to simulate (must be multiple of 10)")
+	rootCmd.AddCommand(resolutionchangeCmd)
+	resolutionchangeCmd.Flags().IntVar(&mpo, "mpo", 10, "how many metrics per org to simulate (must be multiple of 10)")
 }
 
-func buildMetric(name string, set, org int) *schema.MetricData {
+func buildResChangeMetric(name string, i, org int) *schema.MetricData {
 	out := &schema.MetricData{
-		Name:     fmt.Sprintf(name, set),
+		Name:     fmt.Sprintf(name, i),
 		OrgId:    org,
 		Interval: 1,
 		Unit:     "ms",
@@ -83,13 +68,32 @@ func buildMetric(name string, set, org int) *schema.MetricData {
 	out.SetId()
 	return out
 }
-func do(metrics []*schema.MetricData, from, to int64, outs []out.Out) {
-	for ts := from; ts < to; ts++ {
+func runResolutionChange(metrics []*schema.MetricData, from, to int64, outs []out.Out) {
+	ts := from
+	interval := int64(1)
+	for ts <= to {
 		if ts%3600 == 0 {
 			fmt.Println("doing ts", ts)
 		}
+		// every 6 hours, increase resolution
+		if ts%3600*6 == 0 {
+			switch interval {
+			case 1:
+				interval = 10
+			case 10:
+				interval = 60
+			case 60:
+				interval = 300
+			}
+			fmt.Println("changing interval to", interval)
+			for i := range metrics {
+				metrics[i].Interval = int(interval)
+				metrics[i].SetId()
+			}
+		}
+
 		for i := range metrics {
-			metrics[i].Time = int64(ts)
+			metrics[i].Time = ts
 			metrics[i].Value = float64(ts % 10)
 		}
 		for _, out := range outs {
@@ -98,5 +102,6 @@ func do(metrics []*schema.MetricData, from, to int64, outs []out.Out) {
 				log.Error(err.Error())
 			}
 		}
+		ts += interval
 	}
 }
