+++ cmd/mt-fakemetrics/cmd/version.go
@@ -17,6 +17,6 @@ var versionCmd = &cobra.Command{
 	Use:   "version",
 	Short: "Print the version number",
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Printf("fakemetrics %s (built with %s)\n", Version, runtime.Version())
+		fmt.Printf("mt-fakemetrics %s (built with %s)\n", Version, runtime.Version())
 	},
 }
