+++ cmd/mt-fakemetrics/make.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo . && docker build -t mt-fakemetrics .
